backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-socketio==5.6.0
passlib==1.7.4
python-dotenv==0.20.0
influxdb-client==1.22.0
python-jose==1.0.0
